requirements.txt
Django>=1.7.1
django-guardian>=1.2.5
django-registration-redux>=1.1
django-bootstrap3>=5.0.3
django-localflavor>=1.1
